s3-binary-cache-store: Fail gracefully on s3 archived objects
If you turn create a Data Lifecycle Manager for your s3 bucket, old objects will be archived (to Amazon glacier) automatically. This puts you in a state where you need to disable the given binary cache completely.

This PR handles this in the same way as objects that no longer exists in the s3 bucket.